HBX-1649: Move class 'org.hibernate.tool.hbm2x.Cfg2JavaTool' to 'org.hibernate.tool.internal.export.pojo.Cfg2JavaTool'
