+++ orm/src/main/java/org/hibernate/tool/hbm2x/pojo/EntityPOJOClass.java
@@ -33,7 +33,7 @@ import org.hibernate.mapping.Table;
 import org.hibernate.mapping.ToOne;
 import org.hibernate.mapping.UniqueKey;
 import org.hibernate.mapping.Value;
-import org.hibernate.tool.hbm2x.Cfg2JavaTool;
+import org.hibernate.tool.internal.export.pojo.Cfg2JavaTool;
 import org.hibernate.type.ForeignKeyDirection;
 
 public class EntityPOJOClass extends BasicPOJOClass {
