+++ orm/src/main/java/org/hibernate/tool/hbm2x/visitor/JavaTypeFromValueVisitor.java
@@ -10,7 +10,7 @@ import org.hibernate.mapping.Set;
 import org.hibernate.mapping.SimpleValue;
 import org.hibernate.mapping.ToOne;
 import org.hibernate.mapping.Value;
-import org.hibernate.tool.hbm2x.Cfg2JavaTool;
+import org.hibernate.tool.internal.export.pojo.Cfg2JavaTool;
 import org.hibernate.type.CompositeCustomType;
 import org.hibernate.type.CustomType;
 import org.hibernate.type.Type;
