+++ test/nodb/src/test/java/org/hibernate/tool/hbm2x/Hbm2JavaInitializationTest/TestCase.java
@@ -7,9 +7,9 @@ package org.hibernate.tool.hbm2x.Hbm2JavaInitializationTest;
 import org.hibernate.boot.Metadata;
 import org.hibernate.mapping.PersistentClass;
 import org.hibernate.mapping.Property;
-import org.hibernate.tool.hbm2x.Cfg2JavaTool;
 import org.hibernate.tool.hbm2x.pojo.ImportContextImpl;
 import org.hibernate.tool.hbm2x.pojo.POJOClass;
+import org.hibernate.tool.internal.export.pojo.Cfg2JavaTool;
 import org.hibernate.tools.test.util.HibernateUtil;
 import org.junit.Assert;
 import org.junit.Before;
